Added animation

+++ src/co/frostbyte/jge/shaders/Shade.java
@@ -6,5 +6,5 @@ import co.frostbyte.jge.components.Component;
 import java.util.Map;
 
 public abstract class Shade extends Component{
-    public abstract void draw(int[] pixels, Map<Integer, Integer> lights, Point point);
+    public abstract void draw(int[] pixels, Map<Integer, Double> lights, Point point);
 }
